pyproject.toml
[project]
name = "librarian"
version = "0.2.0"
description = "A local-first research scout for coding agents"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastembed>=0.4.0",
    "lancedb>=0.15.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]

[tool.setuptools]
packages = []
